[DS-3439] Copy collection template item specified metadata during Sword v2 METS deposit ingestion.
