+++ dspace-api/src/main/java/org/dspace/content/WorkspaceItemServiceImpl.java
@@ -12,11 +12,8 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.UUID;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
 import org.dspace.app.util.DCInputsReaderException;
 import org.dspace.app.util.Util;
@@ -134,41 +131,7 @@ public class WorkspaceItemServiceImpl implements WorkspaceItemService {
             .addPolicy(context, item, Constants.DELETE, item.getSubmitter(), ResourcePolicy.TYPE_SUBMISSION);
 
         // Copy template if appropriate
-        Item templateItem = collection.getTemplateItem();
-
-        Optional<MetadataValue> colEntityType = getDSpaceEntityType(collection);
-        Optional<MetadataValue> templateItemEntityType = getDSpaceEntityType(templateItem);
-
-        if (template && colEntityType.isPresent() && templateItemEntityType.isPresent() &&
-                !StringUtils.equals(colEntityType.get().getValue(), templateItemEntityType.get().getValue())) {
-            throw new IllegalStateException("The template item has entity type : (" +
-                      templateItemEntityType.get().getValue() + ") different than collection entity type : " +
-                      colEntityType.get().getValue());
-        }
-
-        if (template && colEntityType.isPresent() && templateItemEntityType.isEmpty()) {
-            MetadataValue original = colEntityType.get();
-            MetadataField metadataField = original.getMetadataField();
-            MetadataSchema metadataSchema = metadataField.getMetadataSchema();
-            // NOTE: dspace.entity.type = <blank> does not make sense
-            //       the collection entity type is by default blank when a collection is first created
-            if (StringUtils.isNotBlank(original.getValue())) {
-                itemService.addMetadata(context, item, metadataSchema.getName(), metadataField.getElement(),
-                                        metadataField.getQualifier(), original.getLanguage(), original.getValue());
-            }
-        }
-
-        if (template && (templateItem != null)) {
-            List<MetadataValue> md = itemService.getMetadata(templateItem, Item.ANY, Item.ANY, Item.ANY, Item.ANY);
-
-            for (MetadataValue aMd : md) {
-                MetadataField metadataField = aMd.getMetadataField();
-                MetadataSchema metadataSchema = metadataField.getMetadataSchema();
-                itemService.addMetadata(context, item, metadataSchema.getName(), metadataField.getElement(),
-                                        metadataField.getQualifier(), aMd.getLanguage(),
-                                        aMd.getValue());
-            }
-        }
+        itemService.populateWithTemplateItemMetadata(context, collection, template, item);
 
         itemService.update(context, item);
 
@@ -202,15 +165,6 @@ public class WorkspaceItemServiceImpl implements WorkspaceItemService {
                 itemService.getIdentifiers(context, item)));
 
         return workspaceItem;
-    }
-
-    private Optional<MetadataValue> getDSpaceEntityType(DSpaceObject dSpaceObject) {
-        return Objects.nonNull(dSpaceObject) ? dSpaceObject.getMetadata()
-                                                           .stream()
-                                                           .filter(x -> x.getMetadataField().toString('.')
-                                                                         .equalsIgnoreCase("dspace.entity.type"))
-                                                           .findFirst()
-                                             : Optional.empty();
     }
 
     @Override
